+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 ~~~~~~
 
 - added sign function to the backend that returns the sign of its input (1 or -1)
+- improved readthedocs documentation
 
 0.17.3
 ~~~~~~